+++ src/OpenTl.ClientApi/Settings/ClientSession.cs
@@ -25,7 +25,7 @@
         [AllowNull]
         public byte[] ServerSalt { get; set; }
 
-        public long? UserId { get; set; }
+        public int? UserId { get; set; }
 
         public string ServerAddress { get; set; }
