+++ tests/OpenTl.ClientApi.MtProto.FunctionalTests/Settings/TestSession.cs
@@ -20,7 +20,7 @@
 
         public byte[] ServerSalt { get; set; }
 
-        public long? UserId { get; set; }
+        public int? UserId { get; set; }
 
         public string ServerAddress { get; set; }
